Added new gen_headers.py
