+++ gen_headers/test.py
@@ -18,7 +18,6 @@ along with this program.  If not, see <https://www.gnu.org/licenses/>.
 
 Tests for new gen header functions
 """
-import peewee
 
 from gen_headers import git
 from gen_headers import regex
@@ -27,7 +26,6 @@ from gen_headers import depgraph
 from gen_headers import db
 import os
 import pprint
-import re
 
 def test():
     git.init_repo()
@@ -88,118 +86,3 @@ def test3():
     struct, _ = db.Struct.get_or_create(tag='TheStruct')
     revision, _ = db.Revision.get_or_create(ver_id=ver, struct_id=struct, src="whoa", crc32=db.crc32("whoa"))
     print([rev for rev in struct.revisions])
-
-def write_header(name_prefix, ver, dep_graph, deps):
-    final_output = ""
-    if len(deps) == 0:
-        return final_output, ""
-    ver_string = lambda a: f"{a[0]}_{a[1]}_{a[2]}"
-    ver_name = lambda b: f"{name_prefix}{ver_string(b)}"
-    import_vers = set()
-    name_mappings = {}
-    src_mappings = {}
-    for d in deps:
-        main_v = (d['major'], d['minor'], d['patch'])
-        dep_v = (d["dep_major"], d["dep_minor"], d["dep_patch"])
-        name_mappings[d['tag']] = f"{d['tag']}{ver_string(main_v)}"
-        src_mappings[d['tag']] = d['src']
-        if main_v == dep_v:
-            continue
-        if d['dep_available']:
-            import_vers.add(dep_v)
-            name_mappings[d['dep_tag']] = f"{d['dep_tag']}{ver_string(dep_v)}"
-        else:
-            name_mappings[d['dep_tag']] = "void"
-    for v in sorted(list(import_vers)):
-        final_output += f'#include "{ver_name(v)}.h"\n'
-    for tag in dep_graph.sort():
-        if tag in src_mappings:
-            final_output += '\n' + src_mappings[tag] + '\n'
-    for orig, nname in name_mappings.items():
-        pattern = rf" {orig}(\s|\*)"
-        replacement = rf" {nname}\1"
-        final_output = re.sub(pattern, replacement, final_output)
-    guard = f"{ver_name(ver)}_H".upper()
-    final_output = f"#ifndef {guard}\n#define {guard}\n" + final_output + "\n#endif"
-    return final_output, f"{ver_name(ver)}.h"
-
-def get_version_macro(obj_name, versions, global_ref="blender_ver", global_ref_type="BlenderVersion", class_ref="data_ptr"):
-    ver_string = lambda a: f"{a[0]}_{a[1]}_{a[2]}"
-    output = f"#define {obj_name.upper()}_BASE_RETURN_BODY(A, B, C) \\\n"
-    versions = sorted(list(versions))
-    for i in range(1, len(versions)):
-        current_ver = versions[i - 1]
-        next_ver = versions[i]
-        ref_name = obj_name + ver_string(current_ver)
-        output += (f"    if ({global_ref} < {global_ref_type}::VER_{ver_string(next_ver)}) \\\n"
-        f"        return A ( B reinterpret_cast<{ref_name}*>({class_ref})-> C); \\\n")
-    last_ref = obj_name + ver_string(versions[-1])
-    output += (f"    return A ( B reinterpret_cast<{last_ref}*>({class_ref})-> C); \n"
-    f"#define {obj_name.upper()}_RETURN_REF(T, M)    {obj_name.upper()}_BASE_RETURN_BODY(T, &, M) \n"
-    f"#define {obj_name.upper()}_RETURN_AS(T, M)     {obj_name.upper()}_BASE_RETURN_BODY(T,, M) \n"
-    f"#define {obj_name.upper()}_RETURN(M)           {obj_name.upper()}_BASE_RETURN_BODY(,, M) \n")
-    return output
-
-
-def write_macros(structs, file_name):
-    output = ""
-    for s, vers in structs.items():
-        macro = get_version_macro(s, vers)
-        output += macro + '\n'
-    guard = file_name.replace('.', '_').upper()
-    return f"#ifndef {guard}\n#define {guard}\n\n" + output + "#endif"
-
-
-def test4():
-    git.init_repo()
-    versions = git.available_versions()
-    dna_dir = 'blender/source/blender/makesdna'
-    for v in versions:
-        db_ver = db.get_ver_ref(v)
-        if not db_ver.loaded:
-            print('Checking out blender version...', v)
-            git.checkout_version(v)
-            print("Loading structs...")
-            s_map = regex.load_structs_from_dir(dna_dir)
-            s_defs = types.from_mapping(s_map)
-            print("Writing structs to db...")
-            db.load_revisions_atomic(v, s_defs)
-            db_ver.loaded = True
-            db_ver.save()
-        deps = db.resolve_dependencies(v)
-        graph, all_tags = depgraph.graph_from_dep_query(deps)
-        graph.remove_dangling()
-        if not db_ver.evaluated:
-            print("Evaluating dependencies...", v)
-            unavailable = all_tags - graph.nodes
-            targets = db.Revision.select().join(db.Struct).where((db.Revision.ver_id == db_ver) & (db.Struct.tag << unavailable))
-            db.single_val_update_atomic(db.Revision, {"rev_id": targets}, {"available": 0})
-            db_ver.evaluated = True
-            db_ver.save()
-            deps = db.resolve_dependencies(v)
-        if len(deps) == 0:
-            continue
-        print("Writing header file...", v)
-        header_out, fname = write_header("makesdna_types", v, graph, deps)
-        os.makedirs("output", exist_ok=True)
-        with open(os.path.join("output", fname), 'w') as f:
-            f.write(header_out)
-    macros_out = write_macros(db.get_version_mappings(), "makesdna_mappings.h")
-    with open(os.path.join("output", "makesdna_mappings.h"), 'w') as f:
-            f.write(macros_out)
-        
-
-        
-        
-        
-        
-        
-
-
-test4()
-#db_ver = db.get_ver_ref((2,25,0))
-#unavailable = {"bPoseChannel", "bPose", "bActionChannel"}
-#targets = db.Revision.select().join(db.Struct).where((db.Revision.ver_id == db_ver) & (db.Struct.tag << unavailable))
-#db.single_val_update_atomic(db.Revision, {"rev_id": targets}, {"available": 0})
-
-#print(gen_version_map("Object", [(1,0,0),(2,0,0),(2,2,0),(3,0,0)]))
